预览数据储备

+++ src/main/java/com/me/szzc/pojo/entity/SwapHouse.java
@@ -280,7 +280,7 @@ public class SwapHouse {
     /**
      * 过渡面积
      */
-    private Long interimArea;
+    private BigDecimal interimArea;
 
     /**
      * 过渡单价
